+++ release-notes.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Development
+
+- Bug fix for detecting RTL in string.
+
 ## v3.5.1 - February 10, 2019
 
 - Bug fix - Viewability for regular widgets didn't work for index > 0